+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to the Branch Web SDK will be documented here.
 The Branch Web SDK adheres to [Semantic Versioning](http://semver.org/).
 
+## [VERSION] - unreleased
+
 ## [v1.6.3] - 2015-07-20
 - Fixed an issue that only loaded the first link click data the user ever clicked on
 - Adds the `.first()` method to the Web SDK, from Titanium and Cordova